making route components as server
