+++ app/pcategories/biography/page.js
@@ -1,15 +1,8 @@
-'use client'
-import { useState } from "react";
 import BkCards from "@/components/BkCards/BkCards";
-import {Books} from "@/utils/Constants";
+import { Books } from "@/utils/Constants";
+import Biography from "./Biography";
 
-function Biography(props) {
-  const [showMore, setShowMore] = useState(false);
-
-  const handleShowMore = () => {
-    setShowMore(true);
-  };
-
+export default function page() {
   return (
     <div className="mx-auto max-w-6xl mt-32 text-center p-4">
       <div className="bg-neutral-100 dark:bg-neutral-800 rounded-lg shadow-md p-6 transition-transform hover:scale-105 mx-6">
@@ -17,13 +10,7 @@ function Biography(props) {
         <p className="text-xl font-bold mt-3">
           Discover captivating life stories and memoirs
         </p>
-        <p
-          className={`mt-3 text-lg text-left ${
-            showMore
-              ? "max-h-96 overflow-hidden transition-max-height duration-500 ease"
-              : ""
-          }`}
-        >
+        <p className="mt-3 text-lg text-left max-h-96 overflow-hidden transition-max-height duration-500 ease">
           Discover captivating tales of individuals who have truly embraced life
           to its fullest. Immerse yourself in a diverse collection of books that
           delve into the lives of renowned icons, infamous figures, and
@@ -36,30 +23,16 @@ function Biography(props) {
         </p>
       </div>
       <div className="my-9">
-          <p className="text-xl font-bold">
-            The most talked-about and well-loved Biography & Memoir titles this
-            month.
-          </p>
-          
-          <BkCards bkhead="Trending titles" books={Books} />
-          <BkCards bkhead="Must-Read Biographies of Great Leaders" books={Books} />
-          
-          <div className={`${!showMore && "pt-10"} pb-10`}>
-            {showMore && (
-              <BkCards bkhead="Everything About Biography & Memoir" books={Books} />
-            )}
-            {!showMore && (
-              <button
-                className="block mx-auto bg-blue-500 text-white border-none px-4 py-2 rounded cursor-pointer text-base hover:bg-blue-600 transition-colors duration-300"
-                onClick={handleShowMore}
-              >
-                Show More
-              </button>
-            )}
-          </div>
+        <p className="text-xl font-bold">
+          The most talked-about and well-loved Biography & Memoir titles this
+          month.
+        </p>
+
+        <BkCards bkhead="Trending titles" books={Books} />
+        <BkCards bkhead="Must-Read Biographies of Great Leaders" books={Books} />
+
+        <Biography />
       </div>
     </div>
   );
 }
-
-export default Biography;
