+++ components/About/Feature.js
@@ -1,3 +1,4 @@
+'use client'
 import React from "react";
 import { Carousel } from 'react-responsive-carousel';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
