+++ components/RStar.jsx
@@ -1,0 +1,6 @@
+'use client'
+import ReactStars from "react-stars";
+
+export default function RStar(props) {
+  return <ReactStars {...props} />;
+}
